requirements.txt
asgiref==3.8.1
Django==5.0.6
django-debug-toolbar==4.4.2
django-friendship==1.9.6
django-htmx==1.17.3
django-livereload-server==0.5.1
pillow==10.4.0
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-pool==3.2.2
PuLP==2.8.0
sqlparse==0.5.0
tornado==6.4.1
typing_extensions==4.12.1
tzdata==2024.1